+++ pkg/lvsdk/time.go
@@ -2,11 +2,11 @@ package lvsdk
 
 import "time"
 
-func Millis(ms int64) time.Duration {
+func Millis(ms int) time.Duration {
 	return time.Duration(ms) * time.Millisecond
 }
 
-func Future(ms int64) time.Time {
+func Future(ms int) time.Time {
 	d := Millis(ms)
 	return time.Now().Add(d)
 }
